Create dictionaries_task
Task:
It is necessary to distribute 450 seats among the parties participating in the elections. First, the sum of the votes cast for each party is calculated and the sum of the votes cast for all parties is calculated. This amount is divided by 450. Next, each party receives as many seats in Parliament, which is equal to an integer part of the division of the number of votes for the party in the first private election.
If, after the first round of the distribution of seats, the sum of the number of seats given to the parties is less than 450, the remaining seats are transferred to one party, in descending order of the fractional part of the quotient of the number of votes for the given party to the first electoral quotient. If for two parties these fractional parts are equal, then the party which received the greatest number of votes shall prevail.